requirements.txt
memory-profiler==0.61.0
numpy==1.26.0
Pillow==10.0.1
psutil==5.9.5
